# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Sistem Perwalian Mahasiswa STMIK Bandung",
        "description": "Dokumentasi REST API Enterprise untuk Sistem Perwalian Mahasiswa STMIK Bandung. Mengelola Autentikasi Sanctum, Data Dosen Wali, Data Mahasiswa, Pengajuan dan Review Perwalian, serta Rekapitulasi Laporan.",
        "contact": {
            "name": "Tim Capstone STMIK Bandung",
            "email": "admin@stmikbandung.ac.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Server API Lokal STMIK Bandung"
        }
    ],
    "paths": {
...
        "/dosen": {
            "get": {
                "tags": [
                    "Dosen Wali"
                ],
                "summary": "Daftar Dosen Wali (Paginated & Filtered)",
                "operationId": "5f923b6cf801699b8a37993a4553a52d",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Cari nama/nidn",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Halaman",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mendapatkan data dosen"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Dosen Wali"
                ],
                "summary": "Tambah Dosen Wali Baru",
                "operationId": "9dd2c0f50b65776a4e00cba3096fd7a4",
                "responses": {
                    "201": {
                        "description": "Dosen berhasil dibuat"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/dosen/all-list": {
            "get": {
                "tags": [
                    "Dosen Wali"
                ],
                "summary": "Dropdown Option List Seluruh Dosen Wali",
                "operationId": "892cb9166b9ff6df5e79f605c3146689",
                "responses": {
                    "200": {
                        "description": "Dropdown list dosen"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/dosen/{id}": {
            "get": {
                "tags": [
                    "Dosen Wali"
                ],
                "summary": "Detail Data Dosen Wali",
                "operationId": "f8ba5157844719eb9f31ce81fa43c568",
                "responses": {
                    "200": {
                        "description": "Detail Dosen"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Dosen Wali"
                ],
                "summary": "Update Data Dosen Wali",
                "operationId": "fbe9af76f3f949d7278ed236b9b3d554",
                "responses": {
                    "200": {
                        "description": "Dosen diperbarui"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Dosen Wali"
                ],
                "summary": "Hapus Data Dosen Wali",
                "operationId": "7893d7c7ea3f6d49a7d71880e142c9c6",
                "responses": {
                    "200": {
                        "description": "Dosen dihapus"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/dosen/assign-wali": {
            "post": {
                "tags": [
                    "Dosen Wali"
                ],
                "summary": "Penetapan Bulk Dosen Wali ke Mahasiswa",
                "operationId": "0a37f369a4aff2522b1b72b7aace537f",
                "responses": {
                    "200": {
                        "description": "Penetapan Dosen Wali sukses"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/mahasiswa": {
            "get": {
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Daftar Mahasiswa (Paginated, Search, Filter)",
                "operationId": "9c5bad21852f6d0a70d6c4bb3c9bd0e2",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Cari nama/NIM",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "prodi",
                        "in": "query",
                        "description": "Filter Prodi",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar mahasiswa didapatkan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Tambah Data Mahasiswa Baru",
                "operationId": "d6619ab44f6a08439e0558480ca2875a",
                "responses": {
                    "201": {
                        "description": "Mahasiswa berhasil dibuat"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/mahasiswa/{id}": {
            "get": {
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Detail Data Mahasiswa",
                "operationId": "9ac2da93bbdd7efef189d18a7327c126",
                "responses": {
                    "200": {
                        "description": "Detail Mahasiswa"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Update Data Mahasiswa",
                "operationId": "9d9a1e59776e7f438a8e1335331427e0",
                "responses": {
                    "200": {
                        "description": "Mahasiswa diperbarui"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Hapus Data Mahasiswa",
                "operationId": "6baa06c53cf296b39ecd30f6fa18240a",
                "responses": {
                    "200": {
                        "description": "Mahasiswa dihapus"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/mahasiswa/import": {
            "post": {
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Import Massal Data Mahasiswa",
                "operationId": "1bf3af8fca1449d5c169308b974a5e33",
                "responses": {
                    "200": {
                        "description": "Import data sukses"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token Bearer Sanctum hasil autentikasi login (Contoh: 'Bearer {token}')",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}